Introduce a struct enumerator implementation for ReadOnlyList that can be reused
